feat: Renamed validator files and added permission validator
