+++ src/commands/player/queue.js
@@ -1,6 +1,6 @@
 const logger = require('../../services/logger');
 const { embedOptions, playerOptions } = require('../../config');
-const { notInVoiceChannel } = require('../../utils/validation/voiceChannelValidation');
+const { notInVoiceChannel } = require('../../utils/validation/voiceChannelValidator');
 const { SlashCommandBuilder, EmbedBuilder } = require('discord.js');
 const { useQueue } = require('discord-player');
