+++ src/utils/validation/searchQueryValidator.js
